[GEOT-6759] Remove un-necessary casts from code, add a QA rule to enforce it

+++ modules/extension/xsd/xsd-core/src/test/java/org/geotools/xs/bindings/XSDateTimeStrategyTest.java
@@ -98,7 +98,7 @@ public class XSDateTimeStrategyTest extends TestSchema {
     }
 
     public void testTime() throws Exception {
-        Integer nil = (Integer) null;
+        Integer nil = null;
         java.sql.Time expected;
 
         expected = new java.sql.Time(timestamp(nil, nil, nil, 10, 53, 24));
