+++ modules/plugin/geopkg/src/main/java/org/geotools/geopkg/GeoPkgDataStoreFactory.java
@@ -206,7 +206,7 @@ public class GeoPkgDataStoreFactory extends JDBCDataStoreFactory {
             config.setPragma(SQLiteConfig.Pragma.SYNCHRONOUS, "OFF");
             config.setReadOnly(true);
         }
-        Object map = (Integer) MEMORY_MAP_SIZE.lookUp(params);
+        Object map = MEMORY_MAP_SIZE.lookUp(params);
         if (map instanceof Integer && ((Integer) map) >= 0) {
             int memoryMB = (Integer) map;
             config.setPragma(SQLiteConfig.Pragma.MMAP_SIZE, String.valueOf(memoryMB * 1024 * 1024));
